+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -156,7 +156,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-03T05:20:40.305Z",
+    "2026-06-03T07:30:45.272Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {